cleanup Conv2Image
- drop SUSystem and assume the user calls this as root anyway (die if not)
- don't use /mnt but a temporary directory for mounts
- remove minix support

+++ lib/Conv2Image.pm
@@ -19,6 +19,7 @@ require Exporter;
 
 use strict 'vars';
 use integer;
+use File::Temp;
 
 sub Conv2Image
 {
@@ -30,50 +31,48 @@ sub Conv2Image
 
   ($image, $dir, $fs, $c_k, $c_inodes, $x_k, $x_inodes) = @_;
 
-  $cnt = 1;
+  die "Error: you must be root to build images\n" if $>;
 
-  SUSystem "umount /mnt 2>/dev/null";
+  $cnt = 1;
 
   if($fs eq 'cramfs') {
     $mkcramfs = "/usr/bin/mkcramfs" if -x "/usr/bin/mkcramfs";
     $mkcramfs = "/sbin/mkfs.cramfs" if -x "/sbin/mkfs.cramfs";
     die "$Script: no mkfs.cramfs\n" unless $mkcramfs;
-    SUSystem "rm -f $image";
+    system "rm -f $image";
     system "touch $image $image.log";	# just to ensure the image gets the correct owner
-    SUSystem "sh -c '$mkcramfs $dir $image >$image.log'" and die "$Script: mkfs.cramfs failed";
+    system "sh -c '$mkcramfs $dir $image >$image.log'" and die "$Script: mkfs.cramfs failed";
     $size = -s $image;
     die "$Script: no image?" if $size == 0;
     $name = $image;
     $name =~ s#^.*/##;
     die "$Script: strange image name" if $name eq "";
-    $name .= " " . ($size >> 10);
-    SUSystem "pcramfs '$name' $image" and die "$Script: pcramfs failed";
     return;
   }
   if($fs eq 'squashfs') {
     $mksquashfs = "/usr/bin/mksquashfs" if -x "/usr/bin/mksquashfs";
     $mksquashfs = "/usr/bin/mksquashfs4" if -x "/usr/bin/mksquashfs4";
     die "$Script: no mksquashfs\n" unless $mksquashfs;
-    SUSystem "rm -f $image";
+    system "rm -f $image";
     system "touch $image $image.log";	# just to ensure the image gets the correct owner
-    SUSystem "sh -c '$mksquashfs $dir $image -comp xz -noappend -no-progress >$image.log'" and die "$Script: mksquashfs failed";
+    system "sh -c '$mksquashfs $dir $image -comp xz -noappend -no-progress >$image.log'" and die "$Script: mksquashfs failed";
     $size = -s $image;
     die "$Script: no image?" if $size == 0;
     return;
   }
   elsif($fs eq 'cpio') {
-    SUSystem "rm -f $image";
+    system "rm -f $image";
     system "touch $image";	# just to ensure the image gets the correct owner
-    SUSystem "sh -c '( cd $dir ; find . | cpio --quiet -o -H newc ) >$image'" and die "$Script: cpio failed";
+    system "sh -c '( cd $dir ; find . | cpio --quiet -o -H newc ) >$image'" and die "$Script: cpio failed";
     return;
   }
-  elsif($fs ne 'ext2' && $fs ne 'minix') {
+  elsif($fs ne 'ext2') {
     die "ERRROR: no support for \"$fs\"!\n"
   }
 
   while($cnt <= 2) {
 #    print ">>$c_k, $c_inodes\n";
-    ( $tmp_k, $blk_size, $tmp_inodes ) = $fs eq 'minix' ? MakeMinixImage($image, $c_k, $c_inodes) : MakeExt2Image($image, $c_k, $c_inodes);
+    ( $tmp_k, $blk_size, $tmp_inodes ) = MakeExt2Image($image, $c_k, $c_inodes);
 
     die "$Script: failed to create a $fs fs on \"$image\"" unless $tmp_inodes;
 
@@ -84,25 +83,27 @@ sub Conv2Image
       printf "$Script: created \"%s\": %u kbyte, %u inodes\n", $image, $tmp_k, $tmp_inodes;
     }
 
-    SUSystem "mount -oloop $image /mnt" and die "$Script: mount failed";
+    my $tmp_dir = File::Temp->newdir(TEMPLATE => "/tmp/Conv2Image.XXXXXXXX");
+
+    system "mount -oloop $image $tmp_dir" and die "$Script: mount failed";
 
     # copy everything
-    if(SUSystem "sh -c 'tar --sparse -C $dir -cf - . | tar -C /mnt -xpf -'") {
-      SUSystem "umount /mnt" and warn "$Script: umount failed";
+    if(system "sh -c 'tar --sparse -C $dir -cf - . | tar -C $tmp_dir -xpf -'") {
+      system "umount $tmp_dir" and warn "$Script: umount failed";
       die "$Script: could not add all files to image";
     }
 
     # check the current disk usage
-    for ( `df -Pk /mnt 2>/dev/null` ) {
+    for ( `df -Pk $tmp_dir 2>/dev/null` ) {
       ($blks, $ublks) = ($1, $2) if /^\S+\s+(\d+)\s+(\d+)/;
     }
 
-    for ( `df -Pki /mnt 2>/dev/null` ) {
+    for ( `df -Pki $tmp_dir 2>/dev/null` ) {
       ($inds, $uinds) = ($1, $2) if /^\S+\s+(\d+)\s+(\d+)/;
     }
 
     # unmount it
-    SUSystem "umount /mnt" and die "$Script: umount failed";
+    system "umount $tmp_dir" and die "$Script: umount failed";
 
 #    print "$Script: $image: ${ublks}k/${blks}k used ($uinds/$inds inodes)\n";
